+++ client/src/game/shapes/variants/asset.ts
@@ -14,6 +14,7 @@ import { LayerName } from "../../models/floor";
 import { loadSvgData } from "../../svg";
 import { floorSystem } from "../../systems/floors";
 import { getProperties } from "../../systems/properties/state";
+import { VisionBlock } from "../../systems/properties/types";
 import { TriangulationTarget, visionState } from "../../vision/state";
 import type { SHAPE_TYPE } from "../types";
 
@@ -85,7 +86,7 @@ export class Asset extends BaseRect implements IAsset {
             const svgs = await loadSvgData(`/static/assets/${this.options.svgAsset}`);
             this.svgData = [...map(svgs.values(), (svg) => ({ svg, rp: this.refPoint, paths: undefined }))];
             const props = getProperties(this.id)!;
-            if (props.blocksVision) {
+            if (props.blocksVision !== VisionBlock.No) {
                 if (this.floorId !== undefined) visionState.recalculateVision(this.floorId);
                 visionState.addToTriangulation({ target: TriangulationTarget.VISION, shape: this.id });
             }
@@ -98,8 +99,13 @@ export class Asset extends BaseRect implements IAsset {
         }
     }
 
-    draw(ctx: CanvasRenderingContext2D, customScale?: { center: GlobalPoint; width: number; height: number }): void {
-        super.draw(ctx, customScale);
+    draw(
+        ctx: CanvasRenderingContext2D,
+        lightRevealRender: boolean,
+        customScale?: { center: GlobalPoint; width: number; height: number },
+    ): void {
+        super.draw(ctx, lightRevealRender, customScale);
+
         const center = g2l(this.center);
         const rp = g2l(this.refPoint);
         const ogH = g2lz(this.h);
@@ -108,8 +114,9 @@ export class Asset extends BaseRect implements IAsset {
         const w = customScale ? customScale.width : ogW;
         const deltaH = (ogH - h) / 2;
         const deltaW = (ogW - w) / 2;
+
         if (!this.#loaded) {
-            ctx.fillStyle = FOG_COLOUR;
+            if (!lightRevealRender) ctx.fillStyle = FOG_COLOUR;
             ctx.fillRect(rp.x - center.x, rp.y - center.y, w, h);
         } else {
             try {
@@ -118,7 +125,8 @@ export class Asset extends BaseRect implements IAsset {
                 console.warn(`Shape ${getGlobalId(this.id) ?? "unknown"} could not load the image ${this.src}`);
             }
         }
-        super.drawPost(ctx);
+
+        super.drawPost(ctx, lightRevealRender);
     }
 
     setImage(url: string, sync: boolean): void {