+++ client/src/game/shapes/variants/polygon.ts
@@ -14,6 +14,7 @@ import type { AuraId } from "../../systems/auras/models";
 import { positionState } from "../../systems/position/state";
 import { getProperties } from "../../systems/properties/state";
 import type { ShapeProperties } from "../../systems/properties/state";
+import { VisionBlock } from "../../systems/properties/types";
 import type { TrackerId } from "../../systems/trackers/models";
 import { visionState } from "../../vision/state";
 import { Shape } from "../shape";
@@ -142,8 +143,10 @@ export class Polygon extends Shape implements IShape {
         super.invalidatePoints();
     }
 
-    draw(ctx: CanvasRenderingContext2D): void {
-        super.draw(ctx);
+    draw(ctx: CanvasRenderingContext2D, lightRevealRender: boolean): void {
+        if (lightRevealRender && this.openPolygon) return;
+
+        super.draw(ctx, lightRevealRender);
 
         const center = g2l(this.center);
 
@@ -151,8 +154,10 @@ export class Polygon extends Shape implements IShape {
         ctx.lineJoin = "round";
         const props = getProperties(this.id)!;
 
-        if (props.fillColour === "fog") ctx.fillStyle = FOG_COLOUR;
-        else ctx.fillStyle = props.fillColour;
+        if (!lightRevealRender) {
+            if (props.fillColour === "fog") ctx.fillStyle = FOG_COLOUR;
+            else ctx.fillStyle = props.fillColour;
+        }
 
         ctx.beginPath();
         let localVertex = subtractP(g2l(this.vertices[0]!), center);
@@ -169,16 +174,18 @@ export class Polygon extends Shape implements IShape {
 
         if (!this.openPolygon) ctx.fill();
 
-        for (const [i, c] of props.strokeColour.entries()) {
-            const lw = this.lineWidth[i] ?? this.lineWidth[0]!;
-            ctx.lineWidth = this.ignoreZoomSize ? lw : g2lz(lw);
+        if (!lightRevealRender) {
+            for (const [i, c] of props.strokeColour.entries()) {
+                const lw = this.lineWidth[i] ?? this.lineWidth[0]!;
+                ctx.lineWidth = this.ignoreZoomSize ? lw : g2lz(lw);
 
-            if (c === "fog") ctx.strokeStyle = FOG_COLOUR;
-            else ctx.strokeStyle = c;
-            ctx.stroke();
+                if (c === "fog") ctx.strokeStyle = FOG_COLOUR;
+                else ctx.strokeStyle = c;
+                ctx.stroke();
+            }
         }
 
-        super.drawPost(ctx);
+        super.drawPost(ctx, lightRevealRender);
     }
 
     contains(point: GlobalPoint, nearbyThreshold?: number): boolean {
@@ -269,7 +276,7 @@ export class Polygon extends Shape implements IShape {
             this.layer?.addShape(
                 newPolygon,
                 SyncMode.FULL_SYNC,
-                props.blocksVision ? InvalidationMode.WITH_LIGHT : InvalidationMode.NORMAL,
+                props.blocksVision !== VisionBlock.No ? InvalidationMode.WITH_LIGHT : InvalidationMode.NORMAL,
             );
 
             this.invalidatePoints();
@@ -328,7 +335,7 @@ export class Polygon extends Shape implements IShape {
         if (invalidate) {
             const props = getProperties(this.id)!;
             if (this.floorId !== undefined) {
-                if (props.blocksVision) visionState.recalculateVision(this.floorId);
+                if (props.blocksVision !== VisionBlock.No) visionState.recalculateVision(this.floorId);
                 if (props.blocksMovement) visionState.recalculateMovement(this.floorId);
             }
             if (!this.preventSync) sendShapePositionUpdate([this], false);