+++ client/src/game/shapes/variants/toggleComposite.ts
@@ -18,6 +18,7 @@ import { accessSystem } from "../../systems/access";
 import { auraSystem } from "../../systems/auras";
 import { getProperties } from "../../systems/properties/state";
 import type { ShapeProperties } from "../../systems/properties/state";
+import { VisionBlock } from "../../systems/properties/types";
 import { selectedSystem } from "../../systems/selected";
 import { TriangulationTarget, visionState } from "../../vision/state";
 import { Shape } from "../shape";
@@ -115,7 +116,7 @@ export class ToggleComposite extends Shape implements IToggleComposite {
                 if (props.isToken) accessSystem.removeOwnedToken(variant.id);
                 if (props.blocksMovement)
                     visionState.removeBlocker(TriangulationTarget.MOVEMENT, variant.floorId, variant, true);
-                if (props.blocksVision)
+                if (props.blocksVision !== VisionBlock.No)
                     visionState.removeBlocker(TriangulationTarget.VISION, variant.floorId, variant, true);
                 if (auraSystem.getAll(variant.id, false).length > 0)
                     visionState.removeVisionSources(variant.floorId, variant.id);
@@ -155,7 +156,7 @@ export class ToggleComposite extends Shape implements IToggleComposite {
         if (newVariant.floorId !== undefined) {
             if (props.blocksMovement)
                 visionState.addBlocker(TriangulationTarget.MOVEMENT, newVariant.id, newVariant.floorId, true);
-            if (props.blocksVision)
+            if (props.blocksVision !== VisionBlock.No)
                 visionState.addBlocker(TriangulationTarget.VISION, newVariant.id, newVariant.floorId, true);
 
             for (const au of auraSystem.getAll(newVariant.id, false)) {